Use Astro pools.
