+++ contracts/v2-zapper/wasm/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod astro_pool;
 pub mod contract;
 pub mod lp_pool;
